refactor(breweries): objects -> breweries
