+++ src/main/java/dev/jsinco/brewery/breweries/BreweryPlayer.java
@@ -1,0 +1,4 @@
+package dev.jsinco.brewery.breweries;
+
+public class BreweryPlayer {
+}
