+++ src/main/java/dev/jsinco/brewery/breweries/Destroyable.java
@@ -1,4 +1,4 @@
-package dev.jsinco.brewery.objects;
+package dev.jsinco.brewery.breweries;
 
 public interface Destroyable {
