--- src/main/java/dev/jsinco/brewery/objects/BreweryPlayer.java
@@ -1,4 +1,0 @@
-package dev.jsinco.brewery.objects;
-
-public class BreweryPlayer {
-}
